Fix bug.
The last-date calculation was off. If the last-date happened to be the same weekday as the start of week (eg Sunday), the  was off by 1 date. As a result, for eg, Jan 1926 was rendering only 30 days. Feb 1926 only rendered 27 days.

This commit fixes the bug by checking if the endDate of a month falls on the same day as the start of the week. If true, it calculates upto the next week.